Add shell script
