+++ runShit.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+echo "Hello Asshole"
+py -m pip install keyboard Image pytesseract pyautogui
